Merge updates from Aatos
Now the entire paper source is in file chep09geant4.tex file. Previously
the content was split in an attempt to support both document styles
(revtex4 and jpconf) but this does not seem to be feasible since some
commands used in the paper source are implemented only in the jpconf style.